pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minirsa-breaker"
version = "0.1.0"
description = "Educational RSA visualiser with hardened CI/CD pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyQt6",
  "sympy",
  "pycryptodome"
]

[project.optional-dependencies]
dev = [
  "black==24.8.0",
  "bandit>=1.7.7",
  "hypothesis>=6.100.0",
  "mypy>=1.11.2",
  "mutmut>=2.4.4",
  "pip-audit>=2.7.3",
  "pytest>=8.3.2",
  "pytest-benchmark>=4.0.0",
  "pytest-cov>=5.0.0",
  "pytest-qt>=4.4.0",
  "pytest-randomly>=3.15.0",
  "pytest-xdist>=3.6.1",
  "pytest-icdiff>=0.8",
  "ruff>=0.6.9",
  "types-setuptools",
  "pytest-approvaltests>=0.1.0"
]
gui = [
  "PyQt6"
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["rsa"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "C4", "SIM", "S"]
ignore = ["S311"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["mini_rsa"]
warn_unused_ignores = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["rsa", "PyQt6.*", "Crypto.*"]
ignore_errors = true

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --durations=10 --benchmark-min-rounds=5"
filterwarnings = [
  "error",
  "ignore:.*PyQt6.*:DeprecationWarning"
]
testpaths = ["tests"]
xfail_strict = true
markers = [
  "gui: tests that require Qt GUI (run with xvfb in CI)",
  "performance: performance guardrail benchmarks",
  "fuzz: fuzz-like property tests"
]

[tool.coverage.run]
branch = true
source = ["mini_rsa"]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_covered = true

[tool.hypothesis]
deadline = 500
derandomize = true
print_blob = true
stateful_step_count = 50

[tool.mutmut]
paths_to_mutate = [
  "src/mini_rsa"
]
runner = "python -m pytest --maxfail=1"
backup = false
tests_dir = ["tests"]
also_copy = [
  "rsa.py"
]
